Refactor into centrality package.
