+++ src/main/java/org/gdms/gdmstopology/centrality/GraphFunctionParser.java
@@ -30,7 +30,7 @@
  * For more information, please consult: <http://wwwc.orbisgis.org/> or contact
  * directly: info_at_ orbisgis.org
  */
-package org.gdms.gdmstopology.function;
+package org.gdms.gdmstopology.centrality;
 
 import java.util.Arrays;
 import org.gdms.data.types.Type;
